Funcionando todo, faltaría como mucho hacer el DELETE y UPDATE de Asignatura

+++ src/main/java/com/bosonit/db1Recode/student/application/StudentServiceImpl.java
@@ -2,6 +2,8 @@ package com.bosonit.db1Recode.student.application;
 
 import com.bosonit.db1Recode.Excepciones.EntityNotFoundException;
 import com.bosonit.db1Recode.Excepciones.UnprocessableEntityException;
+import com.bosonit.db1Recode.asignatura.domain.Asignatura;
+import com.bosonit.db1Recode.asignatura.infraestructure.repository.AsignaturaRepository;
 import com.bosonit.db1Recode.student.infraestructure.controller.input.StudentInputDTO;
 import com.bosonit.db1Recode.student.infraestructure.controller.output.StudentOutputDTO;
 import com.bosonit.db1Recode.persona.application.PersonaService;
@@ -13,6 +15,7 @@ import com.bosonit.db1Recode.student.domain.Student;
 import com.bosonit.db1Recode.student.infraestructure.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -32,6 +35,9 @@ public class StudentServiceImpl implements  StudentService{
 
     @Autowired
     ProfesorRepository profesorRepository;
+
+    @Autowired
+    AsignaturaRepository asignaturaRepository;
 
     @Override
     public StudentOutputDTO createStudent(StudentInputDTO studentInputDTO) throws UnprocessableEntityException, EntityNotFoundException {
@@ -88,7 +94,11 @@ public class StudentServiceImpl implements  StudentService{
             throw new EntityNotFoundException("NO HAY NINGUN ESTUDIANTE CREADO AUN", 404);
         }
 
+        return studentList.stream().map(StudentOutputDTO::new).collect(Collectors.toList());
+        /*
+        TODO PREGUNTAR POR QUE ES MEJOR LA LINEA DE ARRIBA QUE LA DE ABAJO:
         return studentList.stream().map(student -> new StudentOutputDTO(student)).collect(Collectors.toList());
+         */
     }
 
     @Override
@@ -142,6 +152,10 @@ public class StudentServiceImpl implements  StudentService{
 
         Student student = studentRepository.findStudentByPersona(persona).orElse(null);
 
+        if (student == null) {
+            throw new EntityNotFoundException("DICHA PERSONA NO ES SALUMNO", 404);
+        }
+
         studentRepository.delete(student);
     }
 
@@ -193,5 +207,33 @@ public class StudentServiceImpl implements  StudentService{
         }
 
         return new StudentOutputDTO(studentRepository.save(student));
+    }
+
+    @Override
+    @Transactional
+    public String setAsignatura(String id_student, List<String> id_asignaturas) throws EntityNotFoundException {
+
+        Student student = studentRepository.findById(id_student).orElse(null);
+
+        if (student == null) {
+            throw new EntityNotFoundException("NO EXISTE NINGUN STUDENT CON DICHO IDSTUDENT", 404);
+        }
+
+        for (String id_asignatura: id_asignaturas) {
+            Asignatura asignatura = asignaturaRepository.findById(id_asignatura).orElse(null);
+
+            if (asignatura == null) {
+                throw new EntityNotFoundException("LA ASIGNATURA CON EL IDASIGNATURA: " + id_asignatura + " NO EXISTE", 404);
+            }
+            if (asignatura.getStudents().contains(student)) {
+                throw new UnprocessableEntityException("ESTE ESTUDIANTE YA TIENE LA ASIGNATURA CON EL ID " + id_asignatura + " asignado", 422);
+            }
+
+            System.err.println("LAS ASIGNATURAS QUE HAY ANTES DE ASIGNARES SON: " + student.getAsignaturaList());
+            asignatura.getStudents().add(student);
+            System.err.println("LAS ASIGNATURAS QUE SE ASIGNAN AL FINAL SON: " + student.getAsignaturaList());
+            asignaturaRepository.save(asignatura);
+        }
+        return "Las asignaturas " + id_asignaturas + " han sido asignadas al estudiante " + id_student;
     }
 }
